moved judger bindings to the dynamo project
