+++ dynamo/src/main/java/io/github/capure/dynamo/judger/JudgerCommunicationException.java
@@ -1,4 +1,4 @@
-package io.github.capure;
+package io.github.capure.dynamo.judger;
 
 public class JudgerCommunicationException extends Exception {
 	public JudgerCommunicationException(String message) {
