+++ dynamo/src/main/java/io/github/capure/dynamo/judger/JudgerErrorException.java
@@ -1,4 +1,4 @@
-package io.github.capure;
+package io.github.capure.dynamo.judger;
 
 public class JudgerErrorException extends Exception {
 	public JudgerErrorException(String message) {
