+++ dynamo/src/main/java/io/github/capure/dynamo/service/CompilerService.java
@@ -1,0 +1,9 @@
+package io.github.capure.dynamo.service;
+
+import io.github.capure.dynamo.config.LangConfig;
+
+public class CompilerService {
+    public void compile(LangConfig config, String sourcePath, String outputPath) {
+        // compile the source code
+    }
+}
