+++ dynamo/src/test/java/io/github/capure/dynamo/judger/JudgerOptionsTest.java
@@ -1,11 +1,11 @@
-package io.github.capure;
+package io.github.capure.dynamo.judger;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class JudgerOptionsTest {
         @Test
