+++ dynamo/src/test/java/io/github/capure/dynamo/judger/JudgerTest.java
@@ -1,6 +1,6 @@
-package io.github.capure;
+package io.github.capure.dynamo.judger;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -8,7 +8,7 @@ import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.List;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class JudgerTest {
 	@Test
